Implementing standard loading component and color
