+++ styles/Colors.js
@@ -1,10 +1,8 @@
 const Colors = {
     primary: '#F18F01',
     secondary: '#3D348B',
+    loading: '#aaaaaa',
     green: '#66B541',
-    light: '#fff',
-    mid: 'green',
-    dark: '#000'
 };
 
 export default Colors;
